+++ Projects/Cubo_do_Nathan/pzcompelpostproc.h
@@ -459,7 +459,7 @@ inline bool TPZCompElPostProc<TCOMPEL>::dataequal(TPZMaterialData &d1,TPZMateria
 		DebugStop();
 		return 0; // dimensions and number of integration points shall match
 	}
-	for(i = 0; i < dim; i++)res += pow(d1.x[i]-d2.x[i],2.); // integration points must be at the same locations
+	for(i = 0; i < dim; i++)res += pow(d1.x[i]-d2.x[i],(REAL)2.); // integration points must be at the same locations
 	/*for(i = 0; i < nshape; i++)
 	{
 		int j;
